+++ spec/integration/integration.spec.js
@@ -74,12 +74,22 @@ describe( 'Integration Test Suite', function() {
 
 				rabbit.addConnection( {
 					name: 'silly',
-					server: 'shfifty-five.gov'
+					server: 'shfifty-five.gov',
+					publishTimeout: 50
 				} );
+
+				rabbit.addExchange( { name: 'silly-ex' }, 'silly' );
 			} );
 
 			it( 'should fail to connect', function() {
 				error.should.equal( 'No endpoints could be reached' );
+			} );
+
+			it( 'should reject publish after timeout', function() {
+				return rabbit.publish( 'silly-ex', { body: 'test' }, 'silly' )
+					.then( null, function( err ) {
+						console.log( err );
+					} );
 			} );
 
 			after( function() {
@@ -439,6 +449,68 @@ describe( 'Integration Test Suite', function() {
 
 		it( 'should receive all messages', function() {
 			harness.received.length.should.equal( messagesToSend );
+		} );
+	} );
+
+	describe( 'with wild card type handling', function() {
+		var harness;
+		before( function( done ) {
+			harness = harnessFn( done, 3 );
+			harness.handle( '#.a' );
+			rabbit.publish( 'wascally-ex.topic', { type: 'one.a', routingKey: 'this.is.one', body: 'one' } );
+			rabbit.publish( 'wascally-ex.topic', { type: 'two.i.a', routingKey: 'this.is.two', body: 'two' } );
+			rabbit.publish( 'wascally-ex.topic', { type: 'three-b.a', routingKey: 'this.is.three', body: 'three' } );
+			rabbit.publish( 'wascally-ex.topic', { type: 'a.four', routingKey: 'this.is.four', body: 'four' } );
+		} );
+
+		it( 'should handle all message types ending in "a"', function() {
+			var results = _.map( harness.received, function( m ) {
+				return {
+					body: m.body,
+					key: m.fields.routingKey
+				};
+			} );
+			_.sortBy( results, 'body' ).should.eql(
+				[
+					{ body: 'one', key: 'this.is.one' },
+					{ body: 'three', key: 'this.is.three' },
+					{ body: 'two', key: 'this.is.two' }
+				] );
+		} );
+
+		it( 'should not handle message types that don\'t match the pattern', function() {
+			harness.unhandled.length.should.equal( 1 );
+			harness.unhandled[ 0 ].body.should.eql( 'four' );
+		} );
+
+		after( function() {
+			harness.clean();
+		} );
+	} );
+
+	describe( 'with no type present', function() {
+		var harness;
+		before( function( done ) {
+			harness = harnessFn( done, 1 );
+			harness.handle( '#.typeless' );
+			rabbit.publish( 'wascally-ex.topic', { type: '', routingKey: 'this.is.typeless', body: 'one' } );
+		} );
+
+		it( 'should handle based on topic', function() {
+			var results = _.map( harness.received, function( m ) {
+				return {
+					body: m.body,
+					key: m.fields.routingKey
+				};
+			} );
+			_.sortBy( results, 'body' ).should.eql(
+				[
+					{ body: 'one', key: 'this.is.typeless' }
+				] );
+		} );
+
+		after( function() {
+			harness.clean();
 		} );
 	} );