+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 0.2.*
 
+### Next
+ * #91 - Provide publish timeout
+ * #98 - No-op duplicate exchange, queue and binding declarations
+ * #96 - Fall back to routing key if no type is provided for routing
+ * #94 - Update amqplib dependnecy to 0.4.0 to support Node 4
+ * #77 - Add explanation of connection events
+ * #95 - Add test coverage for wildcards in handle
+ * Change log namespaces to use '.' delimiter instead of ':'
+
 ### 0.2.8
  * Bug fix - Resolve issue #64 working with node-config no longer throws exceptions by avoiding `get` call
  * Minor improvement to README to better explain publish promises